Laid the groundwork for a SyncAppendList
- Pulled the main struct definition out into its own module
- Pulled index-chunk mapping out into its own module
